Time: 0 ms (100%), Space: 8.6 MB (39.46%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -4,8 +4,8 @@ public:
     if(x<0){
         return false;
     }    
-    int o=x;
-    int r=0;
+    long long o=x;
+    long long r=0;
     while(x>0){
         int d= x%10;
         r= r*10 + d;
